refactor(dm-tool): unify creature detail pane between combat and monster browser
Extract CreatureDetailPane (two-column stat-block layout) shared by both
the monster browser overlay and the combat tracker sidebar. Extract
CombatHpBanner (HP bar + current-turn chrome) from CombatantStatBlock so
combat-specific chrome composes cleanly with the shared content pane.

Move monster-skills.ts/test to features/creatures/ since it is creature
data formatting, not monster-browser-specific logic.
